requirements.txt
PyNaCl>=1.5.0
pytest>=8.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
anyio[trio]>=4.0.0
